Convert EXPECT_UNIT_EQ to macro
Prior to this it was three seperate functions.  Converted it to a macro which calls a class manipulator to class AssertHelper inside of gtest.
Allows the macro to accept a stream of messages like the gtest macros.
A manipulator class acts like a stream and may have a stream inside of it.