Use of EXPECT_UNIT_EQ for tests
The tests which dont use EXPECT_UNIT_EQ are testing the comparison operators in SITest.cpp. The goal is stop using function UnitType::amount() except for the hidden use inside of test.  The goal is to not depend on function amount() where using it is not the true nature of this project.